Refactor coilset opt problems into separate files
